rename auth context permission methods add auth error instead of using ioerror
