+++ src/auth/Cargo.toml
@@ -19,7 +19,6 @@ tracing-futures = "0.2.4"
 futures-util = { version = "0.3.5" }
 
 fluvio-protocol = { version = "0.2.0" }
-fluvio-service = { version = "0.2.0" }
 fluvio-socket = { version = "0.3.0" }
 fluvio-stream-model = { path = "../stream-model", version = "0.2.0" }
 fluvio-sc-schema = { version = "0.1.0", path = "../sc-schema" }
