+++ src/sc/src/error.rs
@@ -5,19 +5,20 @@
 use std::fmt;
 use std::io::Error as IoError;
 
-//use async_channel::SendError;
+
 
 use fluvio_types::PartitionError;
 use k8_client::ClientError;
 use fluvio_socket::FlvSocketError;
+use fluvio_auth::AuthError;
 
 #[derive(Debug)]
 pub enum ScError {
     IoError(IoError),
-    //   SendError(SendError),
     ClientError(ClientError),
     SocketError(FlvSocketError),
     PartitionError(PartitionError),
+    AuthError(AuthError)
 }
 
 impl fmt::Display for ScError {
@@ -28,6 +29,7 @@ impl fmt::Display for ScError {
             Self::ClientError(err) => write!(f, "{}", err),
             Self::SocketError(err) => write!(f, "{}", err),
             Self::PartitionError(err) => write!(f, "{}", err),
+            Self::AuthError(err) => write!(f,"{}",err)
         }
     }
 }
@@ -38,13 +40,13 @@ impl From<IoError> for ScError {
     }
 }
 
-/*
-impl From<SendError> for ScError {
-    fn from(error: SendError) -> Self {
-        Self::SendError(error)
+
+impl From<AuthError> for ScError {
+    fn from(error: AuthError) -> Self {
+        Self::AuthError(error)
     }
 }
-*/
+
 
 impl From<ClientError> for ScError {
     fn from(error: ClientError) -> Self {
